# Datetime parsing optimizations