--- a/vegafusion-core/src/planning/split_domain_data.rs
+++ b/vegafusion-core/src/planning/split_domain_data.rs
@@ -64,159 +64,149 @@ impl<'a> SplitUrlDataNodeVisitor<'a> {
 impl<'a> MutChartVisitor for SplitUrlDataNodeVisitor<'a> {
     fn visit_scale(&mut self, scale: &mut ScaleSpec, scope: &[u32]) -> Result<()> {
         let discrete_scale = scale.type_.clone().unwrap_or_default().is_discrete();
-        if let Some(domain) = &scale.domain {
-            match domain {
-                ScaleDomainSpec::FieldReference(field_ref) => {
-                    let data_name = field_ref.data.clone();
-                    let field_name = &field_ref.field;
+        if let Some(ScaleDomainSpec::FieldReference(field_ref)) = &scale.domain {
+            let data_name = field_ref.data.clone();
+            let field_name = &field_ref.field;
 
-                    // Validate whether we can do anything
-                    if field_name.contains('.') {
-                        // Nested fields not supported
-                        return Ok(());
-                    }
-
-                    // Build suffix using scope
-                    let mut scope_suffix = scope.iter().map(|s| s.to_string()).join("_");
-                    if !scope_suffix.is_empty() {
-                        scope_suffix.insert(0, '_');
-                    }
-
-                    let new_data_name = format!(
-                        "{}_{}_domain_{}{}",
-                        data_name, scale.name, field_name, scope_suffix
-                    );
-                    let (new_data, new_domain) = if discrete_scale {
-                        // Extract sort field and op
-                        let (sort_field, sort_op) =
-                            if let Some(ScaleDataReferenceSort::Parameters(sort_params)) =
-                                &field_ref.sort
-                            {
-                                (sort_params.field.clone(), sort_params.op.clone())
-                            } else {
-                                (None, None)
-                            };
-
-                        let new_data = if let Some(sort_op) = sort_op {
-                            // Will sort by the result of an aggregation operation
-                            let sort_field = sort_field.unwrap_or_else(|| field_name.clone());
-                            serde_json::from_value(serde_json::json!(
-                                {
-                                    "name": new_data_name,
-                                    "source": data_name,
-                                    "transform": [
-                                        {
-                                            "type": "aggregate",
-                                            "as": ["sort_field"],
-                                            "groupby": [field_name],
-                                            "ops": [sort_op],
-                                            "fields": [sort_field]
-                                        }
-                                    ]
-                               }
-                            ))
-                            .unwrap()
-                        } else {
-                            // Will sort by the grouped field values
-                            serde_json::from_value(serde_json::json!(
-                                {
-                                    "name": new_data_name,
-                                    "source": data_name,
-                                    "transform": [
-                                        {
-                                            "type": "aggregate",
-                                            "as": [],
-                                            "groupby": [field_name],
-                                            "ops": [],
-                                            "fields": []
-                                        }, {
-                                            "type": "formula",
-                                            "as": "sort_field",
-                                            "expr": format!("datum['{}']", field_name)
-                                        }
-                                    ]
-                               }
-                            ))
-                            .unwrap()
-                        };
-
-                        // Create new domain specification that uses the new dataset
-                        let sort = match &field_ref.sort {
-                            Some(ScaleDataReferenceSort::Parameters(sort_params)) => {
-                                Some(ScaleDataReferenceSort::Parameters(
-                                    ScaleDataReferenceSortParameters {
-                                        op: Some(AggregateOpSpec::Max),
-                                        field: Some("sort_field".to_string()),
-                                        ..sort_params.clone()
-                                    },
-                                ))
-                            }
-                            sort => sort.clone(),
-                        };
-
-                        let new_domain = ScaleDomainSpec::FieldReference(ScaleDataReferenceSpec {
-                            data: new_data_name,
-                            field: field_name.clone(),
-                            sort,
-                            extra: Default::default(),
-                        });
-
-                        (new_data, new_domain)
-                    } else if matches!(
-                        scale.type_.clone().unwrap_or_default(),
-                        ScaleTypeSpec::Linear
-                    ) {
-                        // Create derived dataset that performs the min/max calculations
-                        let new_data: DataSpec = serde_json::from_value(serde_json::json!(
-                            {
-                                "name": new_data_name,
-                                "source": data_name,
-                                "transform": [
-                                    {
-                                        "type": "formula",
-                                        "as": field_name,
-                                        "expr": format!("+datum['{}']", field_name)
-                                    }, {
-                                        "type": "aggregate",
-                                        "fields": [field_name, field_name],
-                                        "ops": ["min", "max"],
-                                        "as": ["min", "max"],
-                                         "groupby": []
-                                    }
-                                ]
-                            }
-                        ))
-                        .unwrap();
-
-                        // Create new domain specification that uses the new dataset
-                        let new_domain: ScaleDomainSpec =
-                            serde_json::from_value(serde_json::json!([
-                                { "signal": format!("(data('{}')[0] || {{}}).min", new_data_name) },
-                                { "signal": format!("(data('{}')[0] || {{}}).max", new_data_name) }
-                            ]))
-                            .unwrap();
-
-                        (new_data, new_domain)
-                    } else {
-                        // Unsupported scale type
-                        return Ok(());
-                    };
-
-                    // Overwrite scale domain with new domain
-                    scale.domain = Some(new_domain);
-
-                    // Add new dataset at same scope as source dataset
-                    let resolved = self
-                        .task_scope
-                        .resolve_scope(&Variable::new_data(data_name.as_str()), scope)?;
-
-                    // Add new dataset at current scope
-                    self.new_datasets.push((resolved.scope, new_data))
-                }
-                // TODO: handle FieldsReference
-                // ScaleDomainSpec::FieldsReference(_) => {}
-                _ => {}
+            // Validate whether we can do anything
+            if field_name.contains('.') {
+                // Nested fields not supported
+                return Ok(());
             }
+
+            // Build suffix using scope
+            let mut scope_suffix = scope.iter().map(|s| s.to_string()).join("_");
+            if !scope_suffix.is_empty() {
+                scope_suffix.insert(0, '_');
+            }
+
+            let new_data_name = format!(
+                "{}_{}_domain_{}{}",
+                data_name, scale.name, field_name, scope_suffix
+            );
+            let (new_data, new_domain) = if discrete_scale {
+                // Extract sort field and op
+                let (sort_field, sort_op) = if let Some(ScaleDataReferenceSort::Parameters(
+                    sort_params,
+                )) = &field_ref.sort
+                {
+                    (sort_params.field.clone(), sort_params.op.clone())
+                } else {
+                    (None, None)
+                };
+
+                let new_data = if let Some(sort_op) = sort_op {
+                    // Will sort by the result of an aggregation operation
+                    let sort_field = sort_field.unwrap_or_else(|| field_name.clone());
+                    serde_json::from_value(serde_json::json!(
+                        {
+                            "name": new_data_name,
+                            "source": data_name,
+                            "transform": [
+                                {
+                                    "type": "aggregate",
+                                    "as": ["sort_field"],
+                                    "groupby": [field_name],
+                                    "ops": [sort_op],
+                                    "fields": [sort_field]
+                                }
+                            ]
+                       }
+                    ))
+                    .unwrap()
+                } else {
+                    // Will sort by the grouped field values
+                    serde_json::from_value(serde_json::json!(
+                        {
+                            "name": new_data_name,
+                            "source": data_name,
+                            "transform": [
+                                {
+                                    "type": "aggregate",
+                                    "as": [],
+                                    "groupby": [field_name],
+                                    "ops": [],
+                                    "fields": []
+                                }, {
+                                    "type": "formula",
+                                    "as": "sort_field",
+                                    "expr": format!("datum['{}']", field_name)
+                                }
+                            ]
+                       }
+                    ))
+                    .unwrap()
+                };
+
+                // Create new domain specification that uses the new dataset
+                let sort = match &field_ref.sort {
+                    Some(ScaleDataReferenceSort::Parameters(sort_params)) => Some(
+                        ScaleDataReferenceSort::Parameters(ScaleDataReferenceSortParameters {
+                            op: Some(AggregateOpSpec::Max),
+                            field: Some("sort_field".to_string()),
+                            ..sort_params.clone()
+                        }),
+                    ),
+                    sort => sort.clone(),
+                };
+
+                let new_domain = ScaleDomainSpec::FieldReference(ScaleDataReferenceSpec {
+                    data: new_data_name,
+                    field: field_name.clone(),
+                    sort,
+                    extra: Default::default(),
+                });
+
+                (new_data, new_domain)
+            } else if matches!(
+                scale.type_.clone().unwrap_or_default(),
+                ScaleTypeSpec::Linear
+            ) {
+                // Create derived dataset that performs the min/max calculations
+                let new_data: DataSpec = serde_json::from_value(serde_json::json!(
+                    {
+                        "name": new_data_name,
+                        "source": data_name,
+                        "transform": [
+                            {
+                                "type": "formula",
+                                "as": field_name,
+                                "expr": format!("+datum['{}']", field_name)
+                            }, {
+                                "type": "aggregate",
+                                "fields": [field_name, field_name],
+                                "ops": ["min", "max"],
+                                "as": ["min", "max"],
+                                 "groupby": []
+                            }
+                        ]
+                    }
+                ))
+                .unwrap();
+
+                // Create new domain specification that uses the new dataset
+                let new_domain: ScaleDomainSpec = serde_json::from_value(serde_json::json!([
+                    { "signal": format!("(data('{}')[0] || {{}}).min", new_data_name) },
+                    { "signal": format!("(data('{}')[0] || {{}}).max", new_data_name) }
+                ]))
+                .unwrap();
+
+                (new_data, new_domain)
+            } else {
+                // Unsupported scale type
+                return Ok(());
+            };
+
+            // Overwrite scale domain with new domain
+            scale.domain = Some(new_domain);
+
+            // Add new dataset at same scope as source dataset
+            let resolved = self
+                .task_scope
+                .resolve_scope(&Variable::new_data(data_name.as_str()), scope)?;
+
+            // Add new dataset at current scope
+            self.new_datasets.push((resolved.scope, new_data))
         }
         Ok(())
     }