--- a/vegafusion-core/src/planning/watch.rs
+++ b/vegafusion-core/src/planning/watch.rs
@@ -154,7 +154,7 @@ impl ExportUpdate {
         (
             Variable {
                 name: self.name.clone(),
-                namespace: namespace,
+                namespace,
             },
             self.scope.clone(),
         )