This PR contains some optimizations for parsing datetimes from strings.  In turns out that trying several predefined formats with chrono is faster than the custom general parsing function we were always using.  Now a variety of common formats are attempted with chrono first before falling back to the current general parsing logic.

### Benchmarks
```
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 14.6s, or reduce sample count to 30.
flights_crossfilter_local_time                                                                            
                        time:   [144.60 ms 146.45 ms 148.37 ms]
                        change: [-67.134% -66.628% -66.048%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

Benchmarking load_flights_crossfilter_data_local: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.4s, or reduce sample count to 50.
load_flights_crossfilter_data_local                                                                            
                        time:   [93.920 ms 95.776 ms 97.805 ms]
                        change: [-75.460% -74.914% -74.265%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe

Benchmarking load_flights_crossfilter_data_utc: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.6s, or reduce sample count to 50.
load_flights_crossfilter_data_utc                                                                            
                        time:   [93.603 ms 95.166 ms 96.933 ms]
                        change: [-74.708% -74.215% -73.679%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

Benchmarking small-sample/load_flights_crossfilter_data_200k_utc: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 35.1s.
small-sample/load_flights_crossfilter_data_200k_utc                                                                          
                        time:   [3.4559 s 3.9599 s 4.5221 s]
                        change: [-89.217% -87.612% -86.002%] (p = 0.00 < 0.05)
                        Performance has improved.

```